python/requirements.txt
# Основные зависимости
requests>=2.28.0

# Для работы с изображениями
Pillow>=10.0.0

# Для API интеграций
openai>=1.0.0
elevenlabs>=0.2.0

# Дополнительные утилиты
python-dotenv>=0.19.0
